+++ linphone-app/assets/languages/de.ts
@@ -1249,6 +1249,19 @@ Server URL ist nicht konfiguriert.</translation>
     </message>
 </context>
 <context>
+    <name>EmojiPicker</name>
+    <message>
+        <source>emojisLastUsed</source>
+        <extracomment>&apos;Last used&apos; : Emoji title for last used section.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>emojisAll</source>
+        <extracomment>&apos;All&apos; Emoji title for all items.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>EphemeralChatRoom</name>
     <message>
         <source>cancelButton</source>
@@ -3579,6 +3592,11 @@ Klicken Sie hier: &lt;a href=&quot;%1&quot;&gt;%1&lt;/a&gt;
     <message>
         <source>versionCheckTypeNightly</source>
         <extracomment>&apos;Nightly&apos; : Keyword for an option to check the nightly version</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>fontsEmojiChange</source>
+        <extracomment>&apos;Emojis&apos; : Label for changing emojis fonts</extracomment>
         <translation type="unfinished"></translation>
     </message>
 </context>