+++ linphone-app/assets/languages/en.ts
@@ -1249,6 +1249,19 @@ Server URL not configured.</translation>
     </message>
 </context>
 <context>
+    <name>EmojiPicker</name>
+    <message>
+        <source>emojisLastUsed</source>
+        <extracomment>&apos;Last used&apos; : Emoji title for last used section.</extracomment>
+        <translation>Last used</translation>
+    </message>
+    <message>
+        <source>emojisAll</source>
+        <extracomment>&apos;All&apos; Emoji title for all items.</extracomment>
+        <translation>All</translation>
+    </message>
+</context>
+<context>
     <name>EphemeralChatRoom</name>
     <message>
         <source>ephemeralTitle</source>
@@ -3605,6 +3618,11 @@ Click here: &lt;a href=&quot;%1&quot;&gt;%1&lt;/a&gt;
         <source>versionCheckTypeNightly</source>
         <extracomment>&apos;Nightly&apos; : Keyword for an option to check the nightly version</extracomment>
         <translation>Nightly</translation>
+    </message>
+    <message>
+        <source>fontsEmojiChange</source>
+        <extracomment>&apos;Emojis&apos; : Label for changing emojis fonts</extracomment>
+        <translation>Emojis</translation>
     </message>
 </context>
 <context>