+++ linphone-app/assets/languages/zh_CN.ts
@@ -1243,6 +1243,19 @@
     </message>
 </context>
 <context>
+    <name>EmojiPicker</name>
+    <message>
+        <source>emojisLastUsed</source>
+        <extracomment>&apos;Last used&apos; : Emoji title for last used section.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>emojisAll</source>
+        <extracomment>&apos;All&apos; Emoji title for all items.</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+<context>
     <name>EphemeralChatRoom</name>
     <message>
         <source>cancelButton</source>
@@ -3566,6 +3579,11 @@
     <message>
         <source>versionCheckTypeNightly</source>
         <extracomment>&apos;Nightly&apos; : Keyword for an option to check the nightly version</extracomment>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>fontsEmojiChange</source>
+        <extracomment>&apos;Emojis&apos; : Label for changing emojis fonts</extracomment>
         <translation type="unfinished"></translation>
     </message>
 </context>